Add TM1Blob and TM1BlobStorage classes for TM1 file system interaction

+++ airflow_provider_tm1/fs/blob_storage.py
@@ -2,6 +2,7 @@ from fsspec import AbstractFileSystem
 from TM1py import TM1Service
 import io 
 
+
 class TM1Blob(io.BytesIO):
     def __init__(self, buffer: bytes, tm1_service: TM1Service, path: str, mode='rb'):
         super().__init__(buffer)
@@ -15,13 +16,15 @@ class TM1Blob(io.BytesIO):
 
 class TM1BlobStorage(AbstractFileSystem):
     """A file system for TM1 that allows interaction with TM1 objects as files."""
-    
-    def __init__(self, tm1_services: TM1Service, **kwargs):
+    protocol = ('tm1', )
+    def __init__(self, tm1_service: TM1Service, **kwargs):
         super().__init__(**kwargs)
-        self._tm1 = tm1_services
-
+        self._tm1 = tm1_service
+        
     def ls(self, path, **kwargs):
         """List files in a given TM1 path."""
+        if not self._tm1:
+            raise ValueError("TM1Service instance is not registered. Use register_tm1_service() to set it.")
         if path == '/':
             return self._tm1.files.get_all_names()
         
@@ -29,6 +32,8 @@ class TM1BlobStorage(AbstractFileSystem):
 
     def open(self, path, mode='rb', **kwargs):
         """Open a file in TM1."""
+        if not self._tm1:
+            raise ValueError("TM1Service instance is not registered. Use register_tm1_service() to set it.")
         if mode not in ['rb', 'wb']:
             raise ValueError("Mode must be 'rb' or 'wb'.")
         
@@ -43,4 +48,6 @@ class TM1BlobStorage(AbstractFileSystem):
 
     def rm(self, path, **kwargs):
         """Remove a file in TM1."""
+        if not self._tm1:
+            raise ValueError("TM1Service instance is not registered. Use register_tm1_service() to set it.")
         self._tm1.files.delete(path)
